Refactoring and requirements.
